Use the C.UTF-8 locale
By exporting LC_ALL and LANG to C.UTF-8 this avoids potential issues
with ASCII encoding for the environment.
This has been seen with usign the Click library which gives the
following warning:

RuntimeError: Click will abort further execution because Python 3 was
configured to use ASCII as encoding for the environment. Consult
https://click.palletsprojects.com/en/7.x/python3/ for mitigation steps.

This system supports the C.UTF-8 locale which is recommended.
You might be able to resolve your issue by exporting the
following environment variables:

    export LC_ALL=C.UTF-8
    export LANG=C.UTF-8